+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 (Add new notes here)
 
+## v1.2.1
+
+### New
+
+* Support suspend/resume of orchestrations
+ 
+### Updates
+
+* SqlOrchestrationService.WaitForInstanceAsync no longer throws `TimeoutException` - only `OperationCanceledException` (previously could be either, depending on timing)
+
 ## v1.2.0
 
 ### New